Add 3 homework. Attempt 2
