--- simple-chat/src/chats.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./chats.css">
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <title>Chats</title>
-</head>
-<body>
-    <div class="container">
-        <header>
-            <div class="menu">
-                <span class="material-icons">menu</span>
-            </div>
-            <div class="headline">
-                <p>Messenger</p>
-            </div>
-            <div class="search">
-                <span class="material-icons">search</span>
-            </div>
-        </header>
-        <div class="chats">
-            <div class="chat">
-                <div class="photo">
-                    <img src="https://simg.nicepng.com/png/small/356-3568165_blank-profile-picture-female.png"  >
-                </div>
-                <div class="chat-info">
-                    <div class="name-text">
-                        <div class="name">Дженниффер</div>
-                        <!-- <div class="text">Ты куда пропал</div> -->
-                        <div class="text"></div>
-                    </div>
-                    <div class="info">
-                        <div class="time"></div>
-                        <div class="condition"><span class="material-icons"></span></div>
-                    </div>
-                </div>
-            </div>        
-        </div>
-        <div class="create-chat">
-            <span class="material-icons">edit</span>
-        </div>
-    </div>
-    <script type="text/javascript" src="./chats.js"></script>
-</body>
-</html>
