+++ simple-chat/src/index.html
@@ -8,42 +8,42 @@
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     </head>
     <body>
-        <div class="container">
+        <div class="container1" id="page1" > <!-- Контейнер страницы с сообщениями чата -->  
             <div class="header">  <!-- Шапка чата с иконками -->  
-                    <div class="back"> 
-                        <span class="material-icons">
-                            arrow_back
+                <div class="back"> 
+                    <span class="material-icons">
+                        arrow_back
+                    </span>
+                </div>
+                <div class="photo">
+                    <img src="https://simg.nicepng.com/png/small/356-3568165_blank-profile-picture-female.png"  >
+                </div>
+                <div class="info">
+                    <div class="name">
+                        <p >Дженниффер</p>
+                    </div>
+                    <div class="last-visit">
+                        <p>
+                            была 2 часа назад
+                        </p>
+                    </div>
+                </div>
+                <div class="search">
+                    <span class="material-icons">
+                        search
+                    </span>
+                </div>
+                <div class="more">
+                    <button  class="dropbtn">
+                        <span class="material-icons" id="more_vert">
+                            more_vert
                         </span>
-                    </div>
-                    <div class="photo">
-                        <img src="https://simg.nicepng.com/png/small/356-3568165_blank-profile-picture-female.png"  >
-                    </div>
-                    <div class="info">
-                        <div class="name">
-                            <p >Дженниффер</p>
-                        </div>
-                        <div class="last-visit">
-                            <p>
-                                была 2 часа назад
-                            </p>
-                        </div>
-                    </div>
-                    <div class="search">
-                        <span class="material-icons">
-                            search
-                        </span>
-                    </div>
-                    <div class="more">
-                        <button  class="dropbtn">
-                            <span class="material-icons" id="more_vert">
-                                more_vert
-                            </span>
-                        </button>
-                        <div id="myDropdown" class="dropdown-content">
-                          <a href="#">Info</a>
-                          <a href="#">Mute</a>
-                        </div>      
-                    </div>               
+                    </button>
+                    <div id="myDropdown" class="dropdown-content">
+                      <a href="#">Info</a>
+                      <a href="#">Mute</a>
+                    </div>      
+                </div>               
             </div>
             <div class="message"></div> <!-- Блок с сообщениями -->  
             <div class="footer"> <!-- Поле ввода сообщения -->  
@@ -52,6 +52,39 @@
                     <span class="material-icons">attachment</span>
                 </form>
                 
+            </div>
+        </div>
+        <div class="container2" id="page2" style="display: none;"> <!-- Контейнер страницы с чатами -->  
+            <header> <!-- Шапка с иконками --> 
+                <div class="menu">
+                    <span class="material-icons">menu</span>
+                </div>
+                <div class="headline">
+                    <p>Messenger</p>
+                </div>
+                <div class="search">
+                    <span class="material-icons">search</span>
+                </div>
+            </header>
+            <div class="chats"> <!-- Блок с чатами --> 
+                <div class="chat">
+                    <div class="photo">
+                        <img src="https://simg.nicepng.com/png/small/356-3568165_blank-profile-picture-female.png"  >
+                    </div>
+                    <div class="chat-info">
+                        <div class="name-text">
+                            <div class="name">Дженниффер</div>
+                            <div class="text"></div>
+                        </div>
+                        <div class="info">
+                            <div class="time"></div>
+                            <div class="condition"><span class="material-icons"></span></div>
+                        </div>
+                    </div>
+                </div>        
+            </div>
+            <div class="create-chat"> <!-- Блок с парящей кнопкой --> 
+                <span class="material-icons">edit</span>
             </div>
         </div>  
         <script type="text/javascript" src="./index.js"></script>
